Adden support another version arz

+++ BUILD_RUN.py
@@ -32,6 +32,8 @@ print("[INFO] 🗑️ Delete " + LIB_PATH)
 if os.path.exists(LIB_PATH):
     shutil.rmtree(LIB_PATH)
     print("[INFO] ✅ Folder lib/arm64-v8a deleted!")
+else:
+     print("[INFO] ✅ Folder lib/arm64-v8a not found!")
 
 ##################################################################################################################
 
@@ -56,11 +58,22 @@ print("[INFO] ⌚ Connect MonetLoader to GTASAInternal.smali...")
 with open(GTASA_INTERNAL_PATH, "r", encoding="utf-8") as file:
     smali_lines = file.readlines()
 
+check_connect = False
+
 for i, line in enumerate(smali_lines):
-    if 'const-string v1, "samp"' in line and "invoke-static {v1}, Ljava/lang/System;->loadLibrary(Ljava/lang/String;)V" in smali_lines[i + 2]:
-        smali_lines.insert(i + 3, '\n    const-string v1, "monetloader"\n\n    invoke-static {v1}, Ljava/lang/System;->loadLibrary(Ljava/lang/String;)V\n\n')
-        print("[INFO] ✅ Successful connected!")
-        break
+    match1 = re.search(r'const-string (v\d+), "samp"', line)
+    if match1:
+        var_name = match1.group(1)
+        
+        if f"invoke-static {{{var_name}}}, Ljava/lang/System;->loadLibrary(Ljava/lang/String;)V" in smali_lines[i + 2]:
+            smali_lines.insert(i + 4, f'\n    const-string {var_name}, "monetloader"\n\n    invoke-static {{{var_name}}}, Ljava/lang/System;->loadLibrary(Ljava/lang/String;)V\n\n')
+            print("[INFO] ✅ Successful connected MonetLoader!")
+            check_connect = True
+            break
+
+if not check_connect:
+    print("[ERROR] ❌ String 'const-string vX, \"samp\"' not found! (X - random number)")
+    input("\n")
 
 with open(GTASA_INTERNAL_PATH, "w", encoding="utf-8") as file:
     file.writelines(smali_lines)
@@ -93,11 +106,14 @@ print("[INFO] ⌚ Rename \"com.arizona21.game.web\" to \"com.arizona.game\" in A
 manifest_data = manifest_data.replace("com.arizona21.game.web", "com.arizona.game")
 print("[INFO] ✅ Successful renaded!")
 
-print("[INFO] ⌚ Raname app name to \"Arizona RPG\"")
+print("[INFO] ⌚ Raname app name to \"Arizona RPG\"...")
 manifest_data = re.sub(r'android:label="@string/app_name"', 'android:label="Arizona RPG"', manifest_data)
-print("[INFO] ✅ Successful renaded!")
+if 'android:label="Arizona RPG"' in manifest_data:
+    print("[INFO] ✅ Successful renaded!")
+else:
+    print("[ERROR] ❌ String 'android:label=\"@string/app_name\"' not found!")
 
-## connect unity ads activity
+# connect unity ads activity
 new_activities = '''
 <activity android:configChanges="fontScale|keyboard|keyboardHidden|locale|mcc|mnc|navigation|orientation|screenLayout|screenSize|smallestScreenSize|touchscreen|uiMode" android:hardwareAccelerated="true" android:name="com.unity3d.services.ads.adunit.AdUnitActivity" android:theme="@android:style/Theme.NoTitleBar.Fullscreen" />
 <activity android:configChanges="fontScale|keyboard|keyboardHidden|locale|mcc|mnc|navigation|orientation|screenLayout|screenSize|smallestScreenSize|touchscreen|uiMode" android:hardwareAccelerated="true" android:name="com.unity3d.services.ads.adunit.AdUnitTransparentActivity" android:theme="@android:style/Theme.Translucent.NoTitleBar.Fullscreen" />
@@ -109,6 +125,10 @@ manifest_data = re.sub(
     r'\1' + new_activities,
     manifest_data
 )
+
+if not 'com.unity3d.services.ads.adunit' in manifest_data:
+    print("[ERROR] ❌ Unity Ads Activity don't adden!")
+
 with open(MANIFEST_PATH, "w", encoding="utf-8") as file:
     file.write(manifest_data)
 
@@ -128,7 +148,7 @@ if 'arizona rpg' in hud_data:
     with open(HUD_PATH, "w", encoding="utf-8") as file:
         file.write(hud_data)
 else:
-    print("[ERROR] ❌ String not found! No changes made.")
+    print("[ERROR] ❌ Hud-text dont find!")
 
 ##################################################################################################################
 
@@ -142,18 +162,19 @@ matches = [i for i, line in enumerate(smali_lines) if "needUpdateMsg" in line]
 
 insert_index = matches[2]
 smali_lines.insert(insert_index + 2, "    const/4 p1, 0x0\n")
-#const/4 p1, 0x0 додано після третього збігу 'needUpdateMsg'
 
 with open(UPDATE_SERVICE_PATH, "w", encoding="utf-8") as file:
     file.writelines(smali_lines)
 
-print("[INFO] ✅ Client update disabled!")
+print("[INFO] ✅ Client update disabled! (adden \"const/4 p1, 0x0\" after 3/4 \"needUpdateMsg\")")
 
 ##################################################################################################################
 
+# Dont actual, call my java code in MainEntrench
+
 # MAIN_ACTIVITY_PATH = PATH + r"\\" + APK_NAME + r"\smali_classes4\com\arizona\launcher\MainActivity.smali"
 
-# print("[7/13] ⌚ Додаю виклик unpackAssets у " + MAIN_ACTIVITY_PATH)
+# print("[7/13] ⌚ Adding unpackAssets to " + MAIN_ACTIVITY_PATH)
 # with open(MAIN_ACTIVITY_PATH, "r", encoding="utf-8") as file:
 #     smali_data = file.readlines()
 
@@ -168,7 +189,7 @@ print("[INFO] ✅ Client update disabled!")
 
 # with open(MAIN_ACTIVITY_PATH, "w", encoding="utf-8") as file:
 #     file.writelines(smali_data)
-# print("[7/13] ✅ Виклик unpackAssets в onCreate додано")
+# print("[INFO] ✅ Call unpackAssets to onCreate adden!")
 
 ##################################################################################################################
 
@@ -179,31 +200,58 @@ print("[INFO] ⌚ Adding MTG MODS functions (lib installer, toast)...")
 with open(MAIN_ENTRENCH_PATH, "r", encoding="utf-8") as file:
     smali_lines = file.readlines()
 
+check_cpu_abi = False
+check_toast = False
+check_version = False
+
 for i, line in enumerate(smali_lines):
-    if "sget-object v3, Landroid/os/Build;->CPU_ABI:Ljava/lang/String;" in line:
-        smali_lines[i] = '    const-string v3, ""\n'
+    match_cpu_abi = re.search(r'sget-object (v\d+), Landroid/os/Build;->CPU_ABI:Ljava/lang/String;', line)
+    if match_cpu_abi:
+        var_name_2 = match_cpu_abi.group(1)
+        smali_lines[i] = f'    const-string {var_name_2}, ""\n'
         print("[INFO] ✅ Change CPU_ABI")
-    if "invoke-virtual {v0}, Landroid/widget/Toast;->show()V" in line:
-        smali_lines[i] = '    invoke-virtual {v0}, Landroid/widget/Toast;->show()V\n\ninvoke-static {p0, p0}, Lcom/arizona/launcher/AssetExtractor;->unpackAssets(Landroid/app/Activity;Landroid/content/Context;)V\n'
-        print("[INFO] ✅ Connected AssetExtractor")
+        check_cpu_abi = True
+        break
 
-version_pattern = re.compile(r'const-string v3, " v(.+) release_web"')
-version_found = None
+if not check_cpu_abi:
+    print("[ERROR] ❌ CPU_ABI not found!")
 
 for i, line in enumerate(smali_lines):
-    match = version_pattern.search(line)
-    if match:
-        version_found = match.group(1)
-        smali_lines[i] = f'    const-string v3, "[MTG MODS]\\n\u2139\ufe0f ARZ v{version_found} \u2139\ufe0f"\n'
-        print(f"[INFO] ✅ Toast changed to ARZ v{version_found}")
+    match_toast = re.search(r'invoke-virtual {(v\d+)}, Landroid/widget/Toast;->show\(\)V', line)
+    if match_toast:
+        var_name_3 = match_toast.group(1)
+        smali_lines[i] = f'    invoke-virtual {{{var_name_3}}}, Landroid/widget/Toast;->show()V\n\n    invoke-static {{p0, p0}}, Lcom/arizona/launcher/AssetExtractor;->unpackAssets(Landroid/app/Activity;Landroid/content/Context;)V\n'
+        print("[INFO] ✅ Connected AssetExtractor for lib installer")
+        check_toast = True
         break
+
+if not check_toast:
+    print("[ERROR] ❌ Don't connected AssetExtractor for lib installer!")
+    input("\n")
+
+version_pattern = re.compile(r'const-string (v\d+), " v(.+) release_web"')
+version_app = ""
+
+for i, line in enumerate(smali_lines):
+    match_version = version_pattern.search(line)
+    if match_version:
+        var_name_4, version_found = match_version.groups()
+        version_app = "v" + version_found
+        smali_lines[i] = f'    const-string {var_name_4}, "[MTG MODS]\\n\u2139\ufe0f ARZ v{version_found} \u2139\ufe0f"\n'
+        print(f"[INFO] ✅ Toast changed to ARZ v{version_found}")
+        check_version = True
+        break
+
+if not check_version:
+    print("[ERROR] ❌ Version release_web not found!")
+    input("\n")
 
 with open(MAIN_ENTRENCH_PATH, "w", encoding="utf-8") as file:
     file.writelines(smali_lines)
 
 ##################################################################################################################
 
-# OFF UPDATE 
+# SKIP LAUCNCHER UPDATE 
 # DON'T ACTUAL
 
 # MAIN_HANDLER_PATH = PATH + r"\\" + APK_NAME + r"\smali_classes4\com\arizona\launcher\MainEntrench$IncomingHandler.smali"
@@ -215,13 +263,13 @@ with open(MAIN_ENTRENCH_PATH, "w", encoding="utf-8") as file:
 
 # if len(matches) >= 4:
 #     smali_lines[matches[3]] = '    const-string v1, "\u041F\u0440\u043E\u043F\u0443\u0441\u0442\u0438\u0442\u044C"\n'
-#     print("[9/13] ✅ Останній збіг 'sget v1, Lcom/arizona/game/R$string;->exit:I' замінено на 'Пропустити'")
+#     print("[INFO] ✅ Останній збіг 'sget v1, Lcom/arizona/game/R$string;->exit:I' замінено на 'Пропустити'")
 
 # matches = [i for i, line in enumerate(smali_lines) if "invoke-virtual {p1, v1, v2}, Lcom/google/android/material/dialog/MaterialAlertDialogBuilder;->setNegativeButton(ILandroid/content/DialogInterface$OnClickListener;)Lcom/google/android/material/dialog/MaterialAlertDialogBuilder;" in line]
 
 # if len(matches) >= 4:
 #     smali_lines[matches[3]] = '    invoke-virtual {p1, v1, v2}, Lcom/google/android/material/dialog/MaterialAlertDialogBuilder;->setNegativeButton(Ljava/lang/CharSequence;Landroid/content/DialogInterface$OnClickListener;)Lcom/google/android/material/dialog/MaterialAlertDialogBuilder;\n'
-#     print("[9/13] ✅ Метод setNegativeButton виправлено")
+#     print("[INFO] ✅ Метод setNegativeButton виправлено")
 
 # with open(MAIN_HANDLER_PATH, "w", encoding="utf-8") as file:
 #     file.writelines(smali_lines)
@@ -251,7 +299,6 @@ with open(MAIN_ENTRENCH_PATH, "w", encoding="utf-8") as file:
 
 print("[INFO] ⌚ Recompiling APK...")
 subprocess.run(["java", "-jar", APKTOOL_PATH, "b", DECODED_DIR], check=True)
-#subprocess.run(f'java -jar "{APKTOOL_PATH}" b "{DECODED_DIR}"', shell=True)
 print("[INFO] ✅ Recompiling succces!")
 
 ##################################################################################################################
@@ -262,29 +309,33 @@ SIGNED_APK = PATH + r"\MonetLoader.apk"
 
 KEYSTORE_PATH = PATH + r"\key.jks"
 KEY_ALIAS = "key0" ## replace for your sight key data
-KEY_PASS = "Bogdan20MTG!" ## replace for your sight key data
+KEY_PASS = "testkey" ## replace for your sight key pass
 
 print("[INFO] ⌚ Signing APK...")
 
-try:
-    if os.path.exists(SIGNED_APK):
-        print(f"[INFO] 🗑️ Delete old signed apk...")
-        os.remove(SIGNED_APK)
-    
-    subprocess.run([
-        "java", "-jar", APKSIGNER_PATH, "sign",
-        "--ks", KEYSTORE_PATH,
-        "--ks-key-alias", KEY_ALIAS,
-        "--ks-pass", f"pass:{KEY_PASS}",
-        "--key-pass", f"pass:{KEY_PASS}",
-        "--out", SIGNED_APK,
-        UNSIGNED_APK
-    ], check=True)
-    
-    print("[INFO] ✅ Signed successfully!")
-    print(f"[INFO] ℹ️ Your launcher: {SIGNED_APK}")
-except subprocess.CalledProcessError as e:
-    print(f"[ERROR] {e}")
+if os.path.exists(KEYSTORE_PATH):
+    try:
+        if os.path.exists(SIGNED_APK):
+            print(f"[INFO] 🗑️ Delete old signed apk...")
+            os.remove(SIGNED_APK)
+        subprocess.run([
+            "java", "-jar", APKSIGNER_PATH, "sign",
+            "--ks", KEYSTORE_PATH,
+            "--ks-key-alias", KEY_ALIAS,
+            "--ks-pass", f"pass:{KEY_PASS}",
+            "--key-pass", f"pass:{KEY_PASS}",
+            "--out", SIGNED_APK,
+            UNSIGNED_APK
+        ], check=True)
+        
+        print("[INFO] ✅ Signed successfully!")
+        print(f"[INFO] ℹ️ Your launcher {version_app}: {SIGNED_APK}")
+    except subprocess.CalledProcessError as e:
+        print(f"[ERROR] {e}")
+else:
+    print("[ERROR] ❌ Don't signing APK! Not find file key.jks")
+    shutil.move(UNSIGNED_APK, SIGNED_APK)
+    print(f"[INFO] ℹ️ Your no_signed launcher {version_app}: {SIGNED_APK}")
 
 ##################################################################################################################
